✅ Add support for component snapshot testing

+++ appcues/src/main/java/com/appcues/data/model/styling/ComponentShadow.kt
@@ -2,7 +2,7 @@ package com.appcues.data.model.styling
 
 internal data class ComponentShadow(
     val color: ComponentColor,
-    val radius: Int = 0,
-    val x: Int = 0,
-    val y: Int = 0,
+    val radius: Double = 0.0,
+    val x: Double = 0.0,
+    val y: Double = 0.0,
 )
