+++ appcues/src/main/java/com/appcues/data/model/styling/ComponentSize.kt
@@ -1,6 +1,6 @@
 package com.appcues.data.model.styling
 
 internal data class ComponentSize(
-    val width: Int,
-    val height: Int,
+    val width: Double,
+    val height: Double,
 )
