+++ appcues/src/main/java/com/appcues/data/remote/response/step/StepContentResponse.kt
@@ -21,5 +21,5 @@ internal data class StepContentResponse(
     val text: String? = null,
     val content: StepContentResponse? = null,
     val distribution: String? = null,
-    val spacing: Int = 0,
+    val spacing: Double = 0.0,
 )
