+++ appcues/src/main/java/com/appcues/data/remote/response/styling/StyleShadowResponse.kt
@@ -5,7 +5,7 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 internal data class StyleShadowResponse(
     val color: StyleColorResponse,
-    val radius: Int = 0,
-    val x: Int = 0,
-    val y: Int = 0,
+    val radius: Double = 0.0,
+    val x: Double = 0.0,
+    val y: Double = 0.0,
 )
