+++ appcues/src/main/java/com/appcues/ui/primitive/HorizontalStackPrimitive.kt
@@ -58,7 +58,7 @@ private fun RowScope.ItemBox(
     )
 }
 
-private fun ComponentDistribution.toHorizontalArrangement(spacing: Int = 0): Arrangement.Horizontal {
+private fun ComponentDistribution.toHorizontalArrangement(spacing: Double = 0.0): Arrangement.Horizontal {
     return when (this) {
         ComponentDistribution.CENTER -> AppcuesArrangement.spacedCenter(spacing)
         ComponentDistribution.EQUAL -> AppcuesArrangement.spacedEvenly(spacing)
@@ -89,20 +89,20 @@ private val items = arrayListOf(
             id = UUID.randomUUID(),
             text = "Button 1",
             style = ComponentStyle(
-                fontSize = 17,
+                fontSize = 17.0,
                 foregroundColor = ComponentColor(light = 0xFFFFFFFF, dark = 0xFFFFFFFF)
             )
         ),
         style = ComponentStyle(
-            paddingTop = 8,
-            paddingTrailing = 18,
-            paddingBottom = 8,
-            paddingLeading = 18,
+            paddingTop = 8.0,
+            paddingTrailing = 18.0,
+            paddingBottom = 8.0,
+            paddingLeading = 18.0,
             backgroundGradient = arrayListOf(
                 ComponentColor(light = 0xFF5C5CFF, dark = 0xFF5C5CFF),
                 ComponentColor(light = 0xFF8960FF, dark = 0xFF8960FF)
             ),
-            cornerRadius = 6
+            cornerRadius = 6.0
         )
     ),
     TextPrimitive(
@@ -151,9 +151,9 @@ internal fun PreviewTestHorizontalDistributionEqual() {
         HorizontalStackPrimitive(
             id = UUID.randomUUID(),
             distribution = ComponentDistribution.EQUAL,
-            spacing = 10,
+            spacing = 10.0,
             style = ComponentStyle(
-                width = 400,
+                width = 400.0,
                 backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
             ),
             items = items
@@ -168,9 +168,9 @@ internal fun PreviewTestHorizontalDistributionCenter() {
         HorizontalStackPrimitive(
             id = UUID.randomUUID(),
             distribution = ComponentDistribution.CENTER,
-            spacing = 8,
+            spacing = 8.0,
             style = ComponentStyle(
-                width = 400,
+                width = 400.0,
                 backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
             ),
             items = items
@@ -184,17 +184,17 @@ internal fun PreviewTestHorizontalLayout() {
     AppcuesPreviewPrimitive {
         HorizontalStackPrimitive(
             id = UUID.randomUUID(),
-            spacing = 20,
+            spacing = 20.0,
             style = ComponentStyle(
-                width = 500,
-                marginTop = 5,
-                marginLeading = 4,
-                marginBottom = 6,
-                marginTrailing = 3,
-                paddingTop = 6,
-                paddingLeading = 7,
-                paddingBottom = 9,
-                paddingTrailing = 10,
+                width = 500.0,
+                marginTop = 5.0,
+                marginLeading = 4.0,
+                marginBottom = 6.0,
+                marginTrailing = 3.0,
+                paddingTop = 6.0,
+                paddingLeading = 7.0,
+                paddingBottom = 9.0,
+                paddingTrailing = 10.0,
                 backgroundColor = ComponentColor(light = 0xFFCDCDFA, dark = 0xFFCDCDFA)
             ),
             items = items
